More work on tags/branches.
Tearing apart code to get it to work properly. Display of branches with
both CVS and Git looks good - however, the code has been removed to view
only commits on a certain branch (since it was CVS specific). This code
will be rewritten in the individual drivers.

Change supportsFeature() -> hasFeature()

Distinguish between heads (branches) and tags in Git.